% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate Variants Using dbNSFP}
\usage{
annotate_variants(query_data, dbnsfp_file, chunk_size = 1000, workers = 6)
}
\arguments{
\item{query_data}{A data frame containing variant information with columns:
`chr` (chromosome), `start` (start position), `end` (end position), `ref` (reference allele), and `alt` (alternate allele).}

\item{dbnsfp_file}{Path to the bgzipped and indexed dbNSFP file.}

\item{chunk_size}{Number of variants to process per chunk. Default is 1000.}

\item{workers}{Number of cores to use for parallelization. Default is 6.}
}
\value{
A data frame with the original query data and dbNSFP annotations.
}
\description{
This function annotates a set of variants using the dbNSFP database.
}
